% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_of_integration.R
\name{check_missing_insample_values}
\alias{check_missing_insample_values}
\title{Check Missing Values in Sample}
\usage{
check_missing_insample_values(X)
}
\arguments{
\item{X}{A (\eqn{T}x\eqn{N})-matrix of \eqn{N} time series with \eqn{T} observations. Data may also be in a time series format (e.g. \code{ts}, \code{zoo} or \code{xts}) or data frame.}
}
\value{
\eqn{N}-dimensional vector, for each series whether missing values are present (\code{TRUE}) or not (\code{FALSE})
}
\description{
Check Missing Values in Sample
}
\references{
Smeekes, S. and Wilms, I. (2023). bootUR: An R Package for Bootstrap Unit Root Tests. \emph{Journal of Statistical Software}, 106(12), 1-39.
}
