% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FREDQD.R
\docType{data}
\name{FREDQD}
\alias{FREDQD}
\title{FRED-QD Data Set}
\format{A time series object containing 247 macro-economic time series, quarterly observed over 243 time points (from 1959-2019). In the examples, the following two time series will be used

GDPC1: Real Gross Domestic Product (column 1)

T5YFFM:  5-year Treasury constant maturity minus federal funds rate (column 202)}
\usage{
FREDQD
}
\description{
FRED-QD: a Quarterly version of the Federal Reserve Economic Database for Macroeconomic Research. The data set is publicly available and updated in real-time through the FRED database.
}
\examples{
# The latest version of the data set can be imported in R via the commands
# geturl <- url("https://s3.amazonaws.com/files.fred.stlouisfed.org/fred-md/quarterly/current.csv")
# fred <- read.csv(geturl)
}
\references{
McCracken, M. W. and Ng, S. (2016), FRED-MD: A monthly database for macroeconomic research, Journal of Business and Economic Statistics, 34(4), 574-589
}
\keyword{datasets}
