% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{generate_inputs}
\alias{generate_inputs}
\title{Auxiliary Function (not accessible to users) to create all inputs for the unit root test functions.}
\usage{
generate_inputs(y, BSQT_test, iADF_test, level, boot, B, union, p.min, p.max,
  ic, dc, detr, q, l, ic.scale, h.rs, k_DWB, ar_AWB)
}
\arguments{
\item{y}{A T-dimensional vector or (TxN)-matrix with data to be tested for unit roots}

\item{BSQT_test}{Logical indicator whether or not to perform the Bootstrap Quantile Test (TRUE) or not (FALSE).}

\item{iADF_test}{Logical indicator whether or not to perform the individual ADF Tests (TRUE) or not (FALSE).}

\item{level}{Desired significance level of the unit root test. Default is 0.05.}

\item{boot}{String for bootstrap method to be used. Options are MBB - moving blocks bootstrap (Moon and Perron, 2012; Smeekes, 2015), BWB - block wild bootstrap (Shao, 2011), DWB - dependent wild bootstrap (Shao, 2010; Smeekes and Urbain, 2014a; Rho and Shao, 2019), AWB - autoregressive wild bootstrap (Smeekes and Urbain, 2014a; Friedrich, Smeekes and Urbain, 2018), SB - sieve bootstrap (Palm, Smeekes and Urbain, 2008; Smeekes and Urbain, 2014a), SWB - sieve wild boostrap (Smeekes and Taylor, 2014b). Default is MBB.}

\item{B}{Number of bootstrap replications. Default is 9999.}

\item{union}{Logical indicator whether or not to use bootstrap union tests (TRUE) or not (FALSE), see Harvey et al. (2012), Smeekes and Taylor (2012). Default is TRUE.}

\item{p.min}{Minimum lag length in augmented Dickey-Fuller (ADF) regression. Default is 0.}

\item{p.max}{Maximum lag length in ADF regression. Default NULL uses the sample size-based rule 12*(T/100)^(1/4).}

\item{ic}{String for information criterion used to select the lag length in the ADF regression. Options are: AIC, BIC, MAIC, MBIC. Default is MAIC.}

\item{dc}{Numeric vector indicating the deterministic specification. Only relevant if union=FALSE. Options are: 0: no deterministics, 1: intercept only, 2: intercept and trend. Combinations thereof are allowed. Default is 1.}

\item{detr}{String vector indicating the type of detrending to be performed. Only relevant if union=FALSE. Options are OLS and/or QD (typically also called GLS, see Elliott, Rothenberg and Stock, 1996). Default is OLS.}

\item{q}{Numeric vector quantiles to be tested Default is to test each unit sequentially.}

\item{l}{Block length used in MBB and BWB.}

\item{ic.scale}{Logical indicator whether or not to use the rescaled information criteria of Cavaliere et al. (2015) (TRUE) or not (FALSE). Default is TRUE.}

\item{h.rs}{Bandwidth used in rescaled information criteria}

\item{k_DWB}{Kernel used in DWB method}

\item{ar_AWB}{AR parameter used in AWB method.}
}
\description{
Auxiliary Function (not accessible to users) to create all inputs for the unit root test functions.
}
