% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{print.bootUR}
\alias{print.bootUR}
\title{Printing Summary Output for Objects of class bootUR}
\usage{
\method{print}{bootUR}(x, digits = max(3, getOption("digits") - 3), ...)
}
\description{
This function prints summary output for objects of class bootUR (for unit root testing)
}
\keyword{internal}
