% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootURtests.R
\name{check_missing_insample_values}
\alias{check_missing_insample_values}
\title{Check Missing Values in Sample}
\usage{
check_missing_insample_values(X)
}
\arguments{
\item{X}{A (TxN)-matrix of N time series with T observations. Data may also be in a time series format (e.g. \verb{ts}, \verb{zoo} or \verb{xts}).}
}
\value{
N-dimensional vector, for each series whether missing values are present (\verb{TRUE}) or not (\verb{FALSE})
}
\description{
Check Missing Values in Sample
}
