% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootURtests.R
\name{diff_mult}
\alias{diff_mult}
\title{Differences of Multiple Time Series}
\usage{
diff_mult(y, d, keep_NAs = TRUE)
}
\arguments{
\item{y}{A (\eqn{T}x\eqn{N})-matrix of \eqn{N} time series with \eqn{T} observations. Data may also be in a time series format (e.g. \code{ts}, \code{zoo} or \code{xts}) or data frame.}

\item{d}{An \eqn{N}-dimensional vector containing the orders}

\item{keep_NAs}{Logical indicator whether or not to keep the \code{NA} values resulting from differencing at the beginning of the sample. Default is \code{TRUE}. If \code{FALSE}, the entire row containing the \code{NA} values is removed.}
}
\value{
The appropriately differenced data in the same format as the original data.
}
\description{
Performs differencing of multiple time series, with possibly different orders for each time series.
}
