% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootURtests.R
\name{find_nonmissing_subsample}
\alias{find_nonmissing_subsample}
\title{Find Non-Missing Subsamples}
\usage{
find_nonmissing_subsample(X)
}
\arguments{
\item{X}{A (\eqn{T}x\eqn{N})-matrix of \eqn{N} time series with \eqn{T} observations. Data may also be in a time series format (e.g. \verb{ts}, \verb{zoo} or \verb{xts}). Assumes a prior check on missing vaues in-sample has been done.}
}
\value{
A list with the following components:
\describe{
\item{\verb{range}}{(2xN)-dimensional matrix containing the first and last non-missing observation in each column of X.}
\item{\verb{all_equal}}{Logical value indicating whether all series have the same non-missing indices.}
}
}
\description{
Find Non-Missing Subsamples
}
