% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{generate_inputs}
\alias{generate_inputs}
\title{Auxiliary Function (not accessible to users) to create all inputs for the unit root test functions.}
\usage{
generate_inputs(y, BSQT_test, iADF_test, level, boot, B, l, ar_AWB, union,
  p.min, p.max, ic, dc, detr, ic.scale, q, h.rs, k_DWB)
}
\arguments{
\item{y}{A \eqn{T}-dimensional vector or a (\eqn{T} x \eqn{N})-matrix of \eqn{N} time series with \eqn{T} observations to be tested for unit roots. Data may also be in a time series format (e.g. \code{ts}, \code{zoo} or \code{xts}).}

\item{BSQT_test}{Logical indicator whether or not to perform the Bootstrap Quantile Test (\verb{TRUE}) or not (\verb{FALSE}).}

\item{iADF_test}{Logical indicator whether or not to perform the individual ADF Tests (\verb{TRUE}) or not (\verb{FALSE}).}

\item{level}{Desired significance level of the unit root test. Default is 0.05.}

\item{boot}{String for bootstrap method to be used. Options are
\describe{
\item{\verb{"MBB"}}{Moving blocks bootstrap, this is the default;}
\item{\verb{"BWB"}}{Block wild bootstrap;}
\item{\verb{"DWB"}}{Dependent wild bootstrap;}
\item{\verb{"AWB"}}{Autoregressive wild bootstrap;}
\item{\verb{"SB"}}{Sieve bootstrap;}
\item{\verb{"SWB"}}{Sieve wild boostrap.}
}}

\item{B}{Number of bootstrap replications. Default is 9999.}

\item{l}{Desired 'block length' in the bootstrap. For the MBB, BWB and DWB boostrap, this is a genuine block length. For the AWB boostrap, the block length is transformed into an autoregressive parameter via the formula \eqn{0.01^(1/l)}; this can be overwritten by setting \verb{ar_AWB} directly. Default sets the block length as a function of the time series length T, via the rule \eqn{l = 1.75 T^(1/3)}.}

\item{ar_AWB}{Autoregressive parameter used in the AWB bootstrap method (\verb{boot = "AWB"}). Can be used to set the parameter directly rather than via the default link to the block length l.}

\item{union}{Logical indicator whether or not to use bootstrap union tests (\verb{TRUE}) or not (\verb{FALSE}). Default is \verb{TRUE}.}

\item{p.min}{Minimum lag length in the augmented Dickey-Fuller regression. Default is 0.}

\item{p.max}{Maximum lag length in the augmented Dickey-Fuller regression. Default uses the sample size-based rule \eqn{12(T/100)^{1/4}}.}

\item{ic}{String for information criterion used to select the lag length in the augmented Dickey-Fuller regression. Options are: \verb{"AIC"}, \verb{"BIC"}, \verb{"MAIC"}, \verb{"MBIC}. Default is \verb{"MAIC"}.}

\item{dc}{Numeric vector indicating the deterministic specification. Only relevant if \code{union = FALSE}. Options are (combinations of)

\verb{0 } no deterministics;

\verb{1 } intercept only;

\verb{2 } intercept and trend.

For the default union test (\code{union = TRUE}) this is not relevant. If \code{union = FALSE}, the default is adding an intercept (a warning is given).}

\item{detr}{String vector indicating the type of detrending to be performed. Only relevant if \verb{union = FALSE}. Options are: \verb{"OLS"} and/or \verb{"QD"} (typically also called GLS). Default is \verb{"OLS"}.}

\item{ic.scale}{Logical indicator whether or not to use the rescaled information criteria (\verb{TRUE}) or not (\verb{FALSE}). Default is \verb{TRUE}.}

\item{q}{Numeric vector of quantiles to be tested. Default is to test each unit sequentially.}

\item{h.rs}{Bandwidth used in rescaled information criteria.}

\item{k_DWB}{Kernel used in DWB bootstrap method.}
}
\description{
Auxiliary Function (not accessible to users) to create all inputs for the unit root test functions.
}
\seealso{
\code{\link{iADFtest}}, \code{\link{BSQTtest}}, \code{\link{bFDRtest}}
}
\keyword{internal}
