% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{print.mult_htest}
\alias{print.mult_htest}
\title{Printing Summary Output for Objects of class mult_htest}
\usage{
\method{print}{mult_htest}(x, ...)
}
\description{
This function prints summary output for objects of class mult_htest (for multiple hypothesis testing)
}
\keyword{internal}
